feat: middleware로 분기처리 구현(#105)
