+++ src/app/(providers)/profile/[userId]/page.tsx
@@ -7,9 +7,9 @@ const Page = () => {
   return (
     <div className={styles.container}>
       <div className={styles.right}>
-      <div className={styles.title}>
-        <h2>마이페이지</h2>
-      </div>
+        <div className={styles.title}>
+          <h2>마이페이지</h2>
+        </div>
         <Profile />
       </div>
       <div className={styles.left}>
@@ -18,15 +18,5 @@ const Page = () => {
     </div>
   );
 };
-
-// 지원님 코드 => 에러로 인해 주석처리 함.
-// const PrivateCreatePage = () => {
-//   return (
-//     <PrivateRouteWrapper>
-//       <Page />
-//     </PrivateRouteWrapper>
-//   );
-// };
-// export default PrivateCreatePage;
 
 export default Page;
