+++ src/app/(providers)/used-goods/create/page.tsx
@@ -9,12 +9,12 @@ const Page = () => {
   );
 };
 
-const PrivateCreatePage = () => {
-  return (
-    <PrivateRouteWrapper>
-      <Page />
-    </PrivateRouteWrapper>
-  );
-};
+// const PrivateCreatePage = () => {
+//   return (
+//     <PrivateRouteWrapper>
+//       <Page />
+//     </PrivateRouteWrapper>
+//   );
+// };
 
 export default Page;
