+++ src/middleware.ts
@@ -1,6 +1,5 @@
 import { createMiddlewareClient } from '@supabase/auth-helpers-nextjs';
 import { NextResponse } from 'next/server';
-
 import type { NextRequest } from 'next/server';
 import { Database } from './shared/supabase/types/supabase';
 
@@ -11,47 +10,44 @@ export async function middleware(req: NextRequest) {
   const supabase = createMiddlewareClient<Database>({ req, res });
   const { pathname } = req.nextUrl;
 
-  // Refresh session if expired - required for Server Components
+  // 현재 세션 정보 가져오기
   const { data } = await supabase.auth.getSession();
-  console.log('세션이다', data);
 
-  if (data.session !== null && req.nextUrl.pathname.startsWith('/auth')) {
+  // 분기처리 부분 : 멍스타그램 작성/수정, 중고물품 작성/수정 , 마이페이지
+  if (data.session !== null && pathname.startsWith('/auth')) {
     return NextResponse.redirect(new URL('/', req.url));
   }
-
-  if (data.session == null && req.nextUrl.pathname.startsWith('/profile')) {
+  if (data.session === null && pathname.startsWith('/profile')) {
     return NextResponse.redirect(
       new URL('/auth/login?alert=로그인 후 이용가능한 서비스입니다', req.url)
     );
   }
-
-  if (data.session === null && req.nextUrl.pathname.startsWith('/used-goods/create')) {
-    // await supabaseAuth.auth.signOut();1
+  if (data.session === null && pathname.startsWith('/used-goods/create')) {
     return NextResponse.redirect(
       new URL('/auth/login?alert=로그인 후 이용가능한 서비스입니다', req.url)
     );
   }
-
-  // 분기처리 부분 : 멍스타그램 작성/수정, 중고물품 작성/수정 , 마이페이지, api,
-
-  // if (data.session == null && req.nextUrl.pathname.startsWith('/used-goods/create')) {
-  //   await supabaseAuth.auth.signOut();
-  //   return NextResponse.redirect(new URL('/auth/login', req.url));
-  // }
-
-  // if (data.session === null) {
-  //   await supabaseAuth.auth.signOut();
-  // }
+  if (data.session === null && pathname.startsWith('/used-goods/update')) {
+    return NextResponse.redirect(
+      new URL('/auth/login?alert=로그인 후 이용가능한 서비스입니다', req.url)
+    );
+  }
+  if (data.session === null && pathname.startsWith('/mungstagram/create')) {
+    return NextResponse.redirect(
+      new URL('/auth/login?alert=로그인 후 이용가능한 서비스입니다', req.url)
+    );
+  }
 
   return res;
 }
 
-// Ensure the middleware is only called for relevant paths.
 export const config = {
   matcher: [
     '/auth/:path*',
     '/used-goods/create',
-    '/profile'
+    '/profile',
+    '/used-goods/update/:path*',
+    '/mungstagram/create'
     /*
      * Match all request paths except for the ones starting with:
      * - _next/static (static files)
